tendermint: fix secp256k1 verification tests
